RuCaptcha_Task/python-libs/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# pylibs/data_loading/load_noise_and_fine_datasets.py: 6
# pylibs/date_utils.py: 50
Click == 7.0

# pylibs/jpeg_utils.py: 5
PyTurboJPEG == 1.1.5

# pylibs/rect_utils.py: 7,8
Shapely == 1.6.4.post2

# pylibs/lines_segmentator.py: 6
albumentations == 0.2.2

# pylibs/data_loading/data_validators.py: 5
attrs == 19.1.0

# pylibs/.ipynb_checkpoints/text_utils-checkpoint.py: 3
# pylibs/text_utils.py: 7
biopython == 1.73

# pylibs/.ipynb_checkpoints/via_utils-checkpoint.py: 5
# pylibs/data_loading/load_noise_and_fine_datasets.py: 7
# pylibs/via_utils.py: 6
cytoolz == 0.9.0.1

# pylibs/data_loading/.ipynb_checkpoints/loading-checkpoint.py: 8
# pylibs/data_loading/elastic.py: 3
# pylibs/type_utils.py: 5
elasticsearch == 6.3.1

# pylibs/via_utils.py: 8
ilio == 0.4.0

# pylibs/img_utils.py: 3
# pylibs/pyplot_utils.py: 1
# pylibs/vis_utils.py: 9
matplotlib == 3.1.1

# pylibs/minio_utils.py: 5
minio == 4.0.18

# pylibs/cutter.py: 4
# pylibs/data_loading/.ipynb_checkpoints/utils-checkpoint.py: 5
# pylibs/data_loading/utils.py: 7
# pylibs/img_utils.py: 4
# pylibs/jpeg_utils.py: 4
# pylibs/lines_segmentator.py: 4
# pylibs/mask2bbox.py: 4
# pylibs/pandas_utils.py: 4
# pylibs/predictor.py: 5
# pylibs/pyplot_utils.py: 2
# pylibs/rect_utils.py: 6
# pylibs/text_utils.py: 6
# pylibs/type_utils.py: 6
# pylibs/types.py: 3
# pylibs/typex.py: 5
# pylibs/vis_utils.py: 8
numpy == 1.15.4

# pylibs/cutter.py: 3
# pylibs/data_loading/.ipynb_checkpoints/utils-checkpoint.py: 4
# pylibs/data_loading/utils.py: 6
# pylibs/mask2bbox.py: 3
# pylibs/rect_utils.py: 5
# pylibs/type_utils.py: 4
# pylibs/vis_utils.py: 7
opencv-contrib-python == 4.1.0.25

# pylibs/cutter.py: 3
# pylibs/data_loading/.ipynb_checkpoints/utils-checkpoint.py: 4
# pylibs/data_loading/utils.py: 6
# pylibs/mask2bbox.py: 3
# pylibs/rect_utils.py: 5
# pylibs/type_utils.py: 4
# pylibs/vis_utils.py: 7
opencv-contrib-python-headless == 4.1.0.25

# pylibs/cutter.py: 3
# pylibs/data_loading/.ipynb_checkpoints/utils-checkpoint.py: 4
# pylibs/data_loading/utils.py: 6
# pylibs/mask2bbox.py: 3
# pylibs/rect_utils.py: 5
# pylibs/type_utils.py: 4
# pylibs/vis_utils.py: 7
opencv-python == 4.1.0.25

# pylibs/cutter.py: 3
# pylibs/data_loading/.ipynb_checkpoints/utils-checkpoint.py: 4
# pylibs/data_loading/utils.py: 6
# pylibs/mask2bbox.py: 3
# pylibs/rect_utils.py: 5
# pylibs/type_utils.py: 4
# pylibs/vis_utils.py: 7
opencv-python-aarch64 == 3.3.0.1

# pylibs/cutter.py: 3
# pylibs/data_loading/.ipynb_checkpoints/utils-checkpoint.py: 4
# pylibs/data_loading/utils.py: 6
# pylibs/mask2bbox.py: 3
# pylibs/rect_utils.py: 5
# pylibs/type_utils.py: 4
# pylibs/vis_utils.py: 7
opencv-python-headless == 4.1.0.25

# pylibs/.ipynb_checkpoints/via_utils-checkpoint.py: 6
# pylibs/data_loading/.ipynb_checkpoints/loading-checkpoint.py: 9
# pylibs/data_loading/load_noise_and_fine_datasets.py: 8
# pylibs/data_loading/loading.py: 6
# pylibs/data_loading/update.py: 4
# pylibs/data_processing/ranking.py: 5
# pylibs/pandas_utils.py: 5
# pylibs/via_utils.py: 7
pandas == 0.24.2

# pylibs/tensorboard_utils.py: 5
pytablewriter == 0.46.1

# pylibs/data_loading/.ipynb_checkpoints/utils-checkpoint.py: 6
# pylibs/data_loading/utils.py: 8
requests == 2.22.0

# setup.py: 1
setuptools == 41.0.1

# pylibs/data_loading/.ipynb_checkpoints/loading-checkpoint.py: 21
# pylibs/data_loading/loading.py: 18
# pylibs/data_loading/update.py: 7
tensorboardX == 1.7

# pylibs/text_utils.py: 8
toolz == 0.9.0

# pylibs/lines_segmentator.py: 5
# pylibs/resnet50.py: 1,2
# pylibs/tensorboard_utils.py: 6
# pylibs/torch_utils.py: 8
# pylibs/types.py: 4
torch == 1.1.0

# pylibs/resnet50.py: 3
torchvision == 0.3.0

# pylibs/data_loading/.ipynb_checkpoints/loading-checkpoint.py: 16,19
# pylibs/data_loading/elastic.py: 4,5
# pylibs/data_loading/load_noise_and_fine_datasets.py: 9
# pylibs/data_loading/loading.py: 12,15
# pylibs/data_loading/update.py: 5
tqdm == 4.32.1

# pylibs/data_loading/.ipynb_checkpoints/loading-checkpoint.py: 20
# pylibs/data_loading/loading.py: 17
# pylibs/data_loading/update.py: 6
# pylibs/json_utils.py: 1
# pylibs/type_utils.py: 7
ujson == 1.35
cached-property == 1.5.1